Remix.
I don't know who's reading this, or if anyone is.

I doubt even I'll look back at this in the future, except as maybe a reference.

FRC is, without a doubt, the most fun I've ever had in my entire life.

It also taught me a lot, not just about programming, but also about 'people skills.'

Packing all of this into a commit message is probably not the most elegant solution, but it's the only one that would work for me.

Students. Countries. Mentors. Teams. Thank you so much for an amazing season.

See you next year.

https://github.com/FRC5892